+++ packages/web/components/Player/PlayerList.tsx
@@ -5,23 +5,16 @@ import React from 'react';
 
 type Props = {
   players: Player[];
-  showSeasonalXP?: boolean;
 };
 
-export const PlayerList: React.FC<Props> = ({
-  players,
-  showSeasonalXP = false,
-}) => (
+export const PlayerList: React.FC<Props> = ({ players }) => (
   <SimpleGrid
     columns={[1, null, 2, 3]}
     spacing={8}
     autoRows="minmax(35rem, auto)"
   >
     {players.map((player, idx) => (
-      <PlayerTile
-        key={player.profile?.username ?? idx}
-        {...{ player, showSeasonalXP }}
-      />
+      <PlayerTile key={player.profile?.username ?? idx} {...{ player }} />
     ))}
   </SimpleGrid>
 );
